Improve category lock security and UI responsiveness
- Implemented `hasMasterPinFlow` and `getLockedCategoryIdsFlow` in `CategoryLockCrypto` to allow UI components to react to security setting changes.
- Updated `SettingsCategoryLockScreen` and `SettingsSecurityScreen` to use the new flows for real-time preference updates.
- Enhanced `TempFolderCleanupWorker` to recursively delete temporary directories and their contents.
- Improved `AndroidCookieJar` robustnes by filtering out malformed cookies.
- Externalized FlareSolverr error messages to string resources for better localization support.
- Updated JaCoCo tool version to 0.8.14 and refined code coverage configurations, including support for Kotlin source directories and updated build directory APIs.
- Fixed a potential integer overflow in `CategoryLockManager` timeout calculation.
- Updated `jacoco-report` GitHub Action to v1.7.2.

+++ app/src/main/java/eu/kanade/tachiyomi/ui/library/CategoryLockManager.kt
@@ -79,7 +79,7 @@ object CategoryLockManager {
         // 0 means never re-lock during session
         if (timeoutMinutes == 0) return
 
-        val timeoutMillis = timeoutMinutes * 60 * 1000L
+        val timeoutMillis = timeoutMinutes.toLong() * 60 * 1000L
         val currentTime = System.currentTimeMillis()
 
         val categoriesToLock = unlockTimestamps.filter { (_, timestamp) ->
